Compares console output using Symbol equality
When evaluating a command in the console that results in an object with a custom equality method, this line would previously assume the object to compare nicely with a symbol. This change forces the use of the equality method on the symbol instead, making it harder for a custom object to mess up the console.