Re-enable syntax highlighting with prism
This commit adds back both the stylesheet and `prism.js`.

`prism.js` script block has the `data-manual` attribute to prevent any
elements from being automatically highlighted.

Then we add a function to lazyload the style using the
[highlightElement](https://prismjs.com/docs/Prism.html#.highlightElement)
API to only highlight elements as they become visible through scrolling.

The bulk of the code to do the lazy loading taken from https://css-tricks.com/the-complete-guide-to-lazy-loading-images/#method-1-trigger-the-image-load-using-javascript-events