Don't constrain easing function to [0,1] output.
There are legitimate uses of easing functions that output an animation
progress <0 or >1. This allows the animation to overshoot its target or
start with a wind-up that actually moves away from its target for a
brief period. For example, the following easing function does not
display correctly without this change:

    proc { |t| GTK::Geometry.cubic_bezier(t, 0, -1, 2, 1)